add: util - auth

+++ components/Signup/index.jsx
@@ -1,46 +1,62 @@
-import React, { useState, useContext } from 'react';
+import React, { useState } from 'react';
 import Router from 'next/router';
 import Link from 'next/link';
 import axios from 'axios';
-import { useAuth } from '../../util/auth';
 
-const Login = () => {
+const Signup = () => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [handle, setHandle] = useState('');
   const [loading, setLoading] = useState(false);
   const [errors, setErrors] = useState({});
-  const { isAuthenticated, setAuthTrue } = useAuth();
 
   const handleSubmit = (event) => {
     event.preventDefault();
     setLoading(true);
-    const userData = {
+    const newUserData = {
       email: email,
       password: password,
+      confirmPassword: confirmPassword,
+      handle: handle,
     };
     axios
       .post(
-        'https://asia-southeast2-shelf-webapp-de58d.cloudfunctions.net/api/login',
-        userData
+        'https://asia-southeast2-shelf-webapp-de58d.cloudfunctions.net/api/signup',
+        newUserData
       )
       .then((res) => {
         localStorage.setItem('FBIdToken', `Bearer ${res.data.token}`);
+        console.log(res.data);
         setLoading(false);
-        setAuthTrue();
         Router.push('/');
       })
       .catch((err) => {
+        // console.log('ERR');
+        // console.error(err);
         setErrors(err.response.data);
         setLoading(false);
       });
   };
 
   const handleChangeEmail = (event) => {
+    // console.log(event.target.value);
     setEmail(event.target.value);
   };
 
   const handleChangePassword = (event) => {
+    // console.log(event.target.value);
     setPassword(event.target.value);
+  };
+
+  const handleChangeConfirmPassword = (event) => {
+    // console.log(event.target.value);
+    setConfirmPassword(event.target.value);
+  };
+
+  const handleChangeHandle = (event) => {
+    // console.log(event.target.value);
+    setHandle(event.target.value);
   };
 
   return (
@@ -75,18 +91,46 @@ const Login = () => {
         <small id='passwordHelpBlock' className='form-text text-muted'>
           {errors.password}
         </small>
+        <div className='form-group'>
+          <label>Confirm Password</label>
+          <input
+            type='password'
+            className='form-control'
+            id='exampleInputConfirmPassword'
+            placeholder='Confirm Password'
+            onChange={handleChangeConfirmPassword}
+            value={confirmPassword}
+          />
+        </div>
+        <small id='confirmPasswordHelpBlock' className='form-text text-muted'>
+          {errors.confirmPassword}
+        </small>
+        <div className='form-group'>
+          <label>Username</label>
+          <input
+            type='text'
+            className='form-control'
+            id='exampleInputHandle'
+            placeholder='Username'
+            onChange={handleChangeHandle}
+            value={handle}
+          />
+        </div>
+        <small id='passwordHelpBlock' className='form-text text-muted'>
+          {errors.handle}
+        </small>
         <button type='submit' className='btn btn-primary' disabled={loading}>
-          Login
+          Signup
         </button>
         <small id='submitHelpBlock' className='form-text text-muted'>
           {errors.general}
         </small>
         <small>
-          Don't have an account? Sign up <Link href='/signup'>here</Link>
+          Already have an account? Login <Link href='/login'>here.</Link>
         </small>
       </form>
     </div>
   );
 };
 
-export default Login;
+export default Signup;
